Fix error not raised in python 3.
This change the behavior a little. Now int are casted into
numpy.int64. This is needed as in Python3 long have replaced int and
there are no more long! So the int case will be used in Python 3.

Forcing the type isn't a problem as we support only this type for
shape everywhere in Theano.